extract DisplayItem to its own file
